requirements.txt
appdirs==1.4.2
cairocffi==0.8.0
CairoSVG==2.1.1
cffi==1.11.2
cssselect2==0.2.1
defusedxml==0.5.0
Django==1.10.6
django-common-helpers==0.9.1
django-cron==0.5.0
django-easy-pdf==0.1.1
django-jchart==0.4.1
django-social-widgets==0.4.0
django-widget-tweaks==1.4.1
djangorestframework==3.5.4
html5lib==1.0b10
httplib2==0.10.3
olefile==0.44
packaging==16.8
pdfrw==0.4
Pillow==4.3.0
pycparser==2.18
pyparsing==2.1.10
PyPDF2==1.26.0
Pyphen==0.9.4
pytz==2017.2
reportlab==3.4.0
six==1.10.0
tinycss2==0.6.1
WeasyPrint==0.41
webencodings==0.5.1
xhtml2pdf==0.2b1
